make test vector test generic
