+++ frost-ed25519/tests/integration_tests.rs
@@ -318,7 +318,6 @@ async fn check_async_sign_with_dealer() {
 fn check_cocktail_dkg_test_vectors() {
     use rand_core::{CryptoRng, RngCore};
     use sha2::{Digest, Sha512};
-    use std::collections::BTreeMap;
 
     struct CounterDrng {
         seed: Vec<u8>,
@@ -398,254 +397,14 @@ fn check_cocktail_dkg_test_vectors() {
 
     impl CryptoRng for CounterDrng {}
 
-    let file: serde_json::Value =
-        serde_json::from_str(include_str!("helpers/cocktail-dkg-ed25519-sha512.json").trim())
-            .unwrap();
-
-    let seed = hex::decode(file["seed"].as_str().unwrap()).unwrap();
-    let cs_id = file["ciphersuite"].as_str().unwrap().as_bytes().to_vec();
-
-    for vector in file["vectors"].as_array().unwrap().iter() {
-        let n = vector["n"].as_u64().unwrap() as u32;
-        let t = vector["t"].as_u64().unwrap() as u32;
-        let context = hex::decode(vector["context"].as_str().unwrap()).unwrap();
-
-        let static_secret_key_bytes: Vec<Vec<u8>> = vector["config"]["static_secret_keys"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|v| hex::decode(v.as_str().unwrap()).unwrap())
-            .collect();
-
-        let expected_ephemeral_pubs: Vec<Vec<u8>> = vector["round1"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|p| hex::decode(p["ephemeral_public_key"].as_str().unwrap()).unwrap())
-            .collect();
-
-        let expected_group_public_key =
-            hex::decode(vector["group_public_key"].as_str().unwrap()).unwrap();
-
-        let expected_shares: Vec<Vec<u8>> = vector["round2"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|p| hex::decode(p["secret_share"].as_str().unwrap()).unwrap())
-            .collect();
-
-        let expected_verification_shares: Vec<Vec<u8>> = vector["round2"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|p| hex::decode(p["verification_share"].as_str().unwrap()).unwrap())
-            .collect();
-
-        let identifiers: Vec<Identifier> = (1..=n as u16).map(|i| i.try_into().unwrap()).collect();
-
-        let mut static_keys: BTreeMap<Identifier, SigningKey> = BTreeMap::new();
-        let mut participants: BTreeMap<Identifier, VerifyingKey> = BTreeMap::new();
-        for (id, key_bytes) in identifiers.iter().zip(static_secret_key_bytes.iter()) {
-            let sk = SigningKey::deserialize(key_bytes).unwrap();
-            let vk = VerifyingKey::from(&sk);
-            static_keys.insert(*id, sk);
-            participants.insert(*id, vk);
-        }
-
-        let extension = b"";
-
-        // Round 1
-        let mut round1_secret_packages: BTreeMap<
-            Identifier,
-            keys::cocktail_dkg::round1::SecretPackage,
-        > = BTreeMap::new();
-        let mut received_round1_packages: BTreeMap<
-            Identifier,
-            BTreeMap<Identifier, keys::cocktail_dkg::round1::Package>,
-        > = BTreeMap::new();
-
-        for (idx, (&id, sk)) in static_keys.iter().enumerate() {
-            let mut rng = CounterDrng::new(&seed, &cs_id, t, n, (idx + 1) as u32);
-            let (secret_pkg, pkg) = keys::cocktail_dkg::part1(
-                id,
-                n as u16,
-                t as u16,
-                sk,
-                &participants,
-                &context,
-                &BTreeMap::new(),
-                &mut rng,
-            )
-            .unwrap();
-
-            let round1_tv = &vector["round1"][idx];
-
-            assert_eq!(
-                <<Ed25519Sha512 as Ciphersuite>::Group>::serialize(pkg.ephemeral_pub())
-                    .unwrap()
-                    .as_ref(),
-                expected_ephemeral_pubs[idx].as_slice(),
-                "participant {} ephemeral public key mismatch",
-                idx + 1
-            );
-
-            let expected_commitment: Vec<Vec<u8>> = round1_tv["vss_commitment"]
-                .as_array()
-                .unwrap()
-                .iter()
-                .map(|v| hex::decode(v.as_str().unwrap()).unwrap())
-                .collect();
-            assert_eq!(
-                pkg.commitment().serialize().unwrap(),
-                expected_commitment,
-                "participant {} VSS commitment mismatch",
-                idx + 1
-            );
-
-            let expected_enc_shares: Vec<Vec<u8>> = round1_tv["encrypted_shares"]
-                .as_array()
-                .unwrap()
-                .iter()
-                .map(|v| hex::decode(v.as_str().unwrap()).unwrap())
-                .collect();
-            for (j, &receiver_id) in identifiers.iter().enumerate() {
-                let actual = &pkg.encrypted_shares()[&receiver_id];
-                assert_eq!(
-                    actual.as_slice(),
-                    expected_enc_shares[j].as_slice(),
-                    "participant {} encrypted share for receiver {} mismatch",
-                    idx + 1,
-                    j + 1
-                );
-            }
-
-            round1_secret_packages.insert(id, secret_pkg);
-            for &receiver_id in participants.keys() {
-                if receiver_id != id {
-                    received_round1_packages
-                        .entry(receiver_id)
-                        .or_default()
-                        .insert(id, pkg.clone());
-                }
-            }
-        }
-
-        // Round 2
-        let mut round2_secret_packages: BTreeMap<
-            Identifier,
-            keys::cocktail_dkg::round2::SecretPackage,
-        > = BTreeMap::new();
-        let mut received_round2_packages: BTreeMap<
-            Identifier,
-            BTreeMap<Identifier, keys::cocktail_dkg::round2::Package>,
-        > = BTreeMap::new();
-
-        for (&id, sk) in static_keys.iter() {
-            let secret_pkg = round1_secret_packages.remove(&id).unwrap();
-            let round1_packages = &received_round1_packages[&id];
-            let (r2_secret, r2_pkg, _received_payloads) = keys::cocktail_dkg::part2(
-                secret_pkg,
-                round1_packages,
-                sk,
-                &participants,
-                &context,
-                extension,
-                rand::rngs::OsRng,
-            )
-            .unwrap();
-
-            round2_secret_packages.insert(id, r2_secret);
-            for &receiver_id in participants.keys() {
-                received_round2_packages
-                    .entry(receiver_id)
-                    .or_default()
-                    .insert(id, r2_pkg.clone());
-            }
-        }
-
-        // Round 3
-        let mut transcript_for_recovery: Vec<u8> = Vec::new();
-        let mut cert_for_recovery: BTreeMap<Identifier, Signature> = BTreeMap::new();
-
-        for (idx, (&id, _)) in static_keys.iter().enumerate() {
-            let r2_secret = &round2_secret_packages[&id];
-            let round2_packages = &received_round2_packages[&id];
-            let (key_pkg, pubkey_pkg, transcript, cert) =
-                keys::cocktail_dkg::part3(r2_secret, round2_packages).unwrap();
-
-            if idx == 0 {
-                transcript_for_recovery = transcript;
-                cert_for_recovery = cert;
-            }
-
-            assert_eq!(
-                pubkey_pkg.verifying_key().serialize().unwrap().as_slice(),
-                expected_group_public_key.as_slice(),
-                "participant {} group public key mismatch",
-                idx + 1
-            );
-            assert_eq!(
-                key_pkg.signing_share().serialize().as_slice(),
-                expected_shares[idx].as_slice(),
-                "participant {} secret share mismatch",
-                idx + 1
-            );
-            assert_eq!(
-                pubkey_pkg
-                    .verifying_shares()
-                    .get(&id)
-                    .unwrap()
-                    .serialize()
-                    .unwrap()
-                    .as_slice(),
-                expected_verification_shares[idx].as_slice(),
-                "participant {} verification share mismatch",
-                idx + 1
-            );
-        }
-
-        // Recovery
-        if let Some(recovery) = vector.get("recovery") {
-            let recovery_id = recovery["participant_id"].as_u64().unwrap() as u16;
-            let recovery_identifier = Identifier::try_from(recovery_id).unwrap();
-            let recovery_sk = static_keys.get(&recovery_identifier).unwrap();
-
-            let ciphertexts_json = recovery["ciphertexts"].as_array().unwrap();
-            let mut recovery_ciphertexts: BTreeMap<Identifier, Vec<u8>> = BTreeMap::new();
-            for (j_idx, ct) in ciphertexts_json.iter().enumerate() {
-                let sender_id = Identifier::try_from((j_idx + 1) as u16).unwrap();
-                recovery_ciphertexts.insert(sender_id, hex::decode(ct.as_str().unwrap()).unwrap());
-            }
-
-            let expected_recovered_share =
-                hex::decode(recovery["recovered_secret_share"].as_str().unwrap()).unwrap();
-            let expected_recovered_vshare =
-                hex::decode(recovery["recovered_verification_share"].as_str().unwrap()).unwrap();
-
-            let (recovered_key_pkg, recovered_pubkey_pkg) = keys::cocktail_dkg::recover(
-                recovery_sk,
-                &transcript_for_recovery,
-                &cert_for_recovery,
-                &recovery_ciphertexts,
-            )
-            .unwrap();
-
-            assert_eq!(
-                recovered_key_pkg.signing_share().serialize().as_slice(),
-                expected_recovered_share.as_slice(),
-                "recovered secret share mismatch"
-            );
-            assert_eq!(
-                recovered_pubkey_pkg
-                    .verifying_shares()
-                    .get(&recovery_identifier)
-                    .unwrap()
-                    .serialize()
-                    .unwrap()
-                    .as_slice(),
-                expected_recovered_vshare.as_slice(),
-                "recovered verification share mismatch"
-            );
-        }
-    }
+    frost_core::tests::ciphersuite_generic::check_cocktail_dkg_test_vectors::<
+        Ed25519Sha512,
+        _,
+        _,
+    >(
+        include_str!("helpers/cocktail-dkg-ed25519-sha512.json"),
+        |seed, cs_id, t, n, p| CounterDrng::new(seed, cs_id, t, n, p),
+        true, // encrypted shares match (XChaCha20Poly1305)
+        true, // recovery is tested
+    );
 }
