+++ frost-ed448/tests/integration_tests.rs
@@ -321,7 +321,6 @@ fn check_cocktail_dkg_test_vectors() {
         digest::{ExtendableOutput, Update, XofReader},
         Shake256,
     };
-    use std::collections::BTreeMap;
 
     struct CounterDrng {
         seed: Vec<u8>,
@@ -400,198 +399,14 @@ fn check_cocktail_dkg_test_vectors() {
 
     impl CryptoRng for CounterDrng {}
 
-    let file: serde_json::Value =
-        serde_json::from_str(include_str!("helpers/cocktail-dkg-ed448-shake256.json").trim())
-            .unwrap();
-
-    let seed = hex::decode(file["seed"].as_str().unwrap()).unwrap();
-    let cs_id = file["ciphersuite"].as_str().unwrap().as_bytes().to_vec();
-
-    for vector in file["vectors"].as_array().unwrap().iter() {
-        let n = vector["n"].as_u64().unwrap() as u32;
-        let t = vector["t"].as_u64().unwrap() as u32;
-        let context = hex::decode(vector["context"].as_str().unwrap()).unwrap();
-
-        let static_secret_key_bytes: Vec<Vec<u8>> = vector["config"]["static_secret_keys"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|v| hex::decode(v.as_str().unwrap()).unwrap())
-            .collect();
-
-        let expected_ephemeral_pubs: Vec<Vec<u8>> = vector["round1"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|p| hex::decode(p["ephemeral_public_key"].as_str().unwrap()).unwrap())
-            .collect();
-
-        let expected_group_public_key =
-            hex::decode(vector["group_public_key"].as_str().unwrap()).unwrap();
-
-        let expected_shares: Vec<Vec<u8>> = vector["round2"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|p| hex::decode(p["secret_share"].as_str().unwrap()).unwrap())
-            .collect();
-
-        let expected_verification_shares: Vec<Vec<u8>> = vector["round2"]
-            .as_array()
-            .unwrap()
-            .iter()
-            .map(|p| hex::decode(p["verification_share"].as_str().unwrap()).unwrap())
-            .collect();
-
-        let identifiers: Vec<Identifier> = (1..=n as u16).map(|i| i.try_into().unwrap()).collect();
-
-        let mut static_keys: BTreeMap<Identifier, SigningKey> = BTreeMap::new();
-        let mut participants: BTreeMap<Identifier, VerifyingKey> = BTreeMap::new();
-        for (id, key_bytes) in identifiers.iter().zip(static_secret_key_bytes.iter()) {
-            // JSON stores 56-byte raw scalars; ed448 expects 57-byte RFC 8032 format (trailing 0x00).
-            let mut key_padded = [0u8; 57];
-            key_padded[..key_bytes.len()].copy_from_slice(key_bytes);
-            let sk = SigningKey::deserialize(&key_padded).unwrap();
-            let vk = VerifyingKey::from(&sk);
-            static_keys.insert(*id, sk);
-            participants.insert(*id, vk);
-        }
-
-        let extension = b"";
-
-        // Round 1
-        let mut round1_secret_packages: BTreeMap<
-            Identifier,
-            keys::cocktail_dkg::round1::SecretPackage,
-        > = BTreeMap::new();
-        let mut received_round1_packages: BTreeMap<
-            Identifier,
-            BTreeMap<Identifier, keys::cocktail_dkg::round1::Package>,
-        > = BTreeMap::new();
-
-        for (idx, (&id, sk)) in static_keys.iter().enumerate() {
-            let mut rng = CounterDrng::new(&seed, &cs_id, t, n, (idx + 1) as u32);
-            let (secret_pkg, pkg) = keys::cocktail_dkg::part1(
-                id,
-                n as u16,
-                t as u16,
-                sk,
-                &participants,
-                &context,
-                &BTreeMap::new(),
-                &mut rng,
-            )
-            .unwrap();
-
-            let round1_tv = &vector["round1"][idx];
-
-            assert_eq!(
-                <<Ed448Shake256 as Ciphersuite>::Group>::serialize(pkg.ephemeral_pub())
-                    .unwrap()
-                    .as_ref(),
-                expected_ephemeral_pubs[idx].as_slice(),
-                "participant {} ephemeral public key mismatch",
-                idx + 1
-            );
-
-            let expected_commitment: Vec<Vec<u8>> = round1_tv["vss_commitment"]
-                .as_array()
-                .unwrap()
-                .iter()
-                .map(|v| hex::decode(v.as_str().unwrap()).unwrap())
-                .collect();
-            assert_eq!(
-                pkg.commitment().serialize().unwrap(),
-                expected_commitment,
-                "participant {} VSS commitment mismatch",
-                idx + 1
-            );
-
-            // Note: encrypted shares are 73 bytes in our implementation (57-byte RFC 8032 scalar +
-            // 16-byte AEAD tag) but 72 bytes in the JSON (56-byte raw scalar + 16-byte AEAD tag).
-            // This format difference means the ciphertexts cannot be compared against the JSON.
-
-            round1_secret_packages.insert(id, secret_pkg);
-            for &receiver_id in participants.keys() {
-                if receiver_id != id {
-                    received_round1_packages
-                        .entry(receiver_id)
-                        .or_default()
-                        .insert(id, pkg.clone());
-                }
-            }
-        }
-
-        // Round 2
-        let mut round2_secret_packages: BTreeMap<
-            Identifier,
-            keys::cocktail_dkg::round2::SecretPackage,
-        > = BTreeMap::new();
-        let mut received_round2_packages: BTreeMap<
-            Identifier,
-            BTreeMap<Identifier, keys::cocktail_dkg::round2::Package>,
-        > = BTreeMap::new();
-
-        for (&id, sk) in static_keys.iter() {
-            let secret_pkg = round1_secret_packages.remove(&id).unwrap();
-            let round1_packages = &received_round1_packages[&id];
-            let (r2_secret, r2_pkg, _received_payloads) = keys::cocktail_dkg::part2(
-                secret_pkg,
-                round1_packages,
-                sk,
-                &participants,
-                &context,
-                extension,
-                rand::rngs::OsRng,
-            )
-            .unwrap();
-
-            round2_secret_packages.insert(id, r2_secret);
-            for &receiver_id in participants.keys() {
-                received_round2_packages
-                    .entry(receiver_id)
-                    .or_default()
-                    .insert(id, r2_pkg.clone());
-            }
-        }
-
-        // Round 3
-        for (idx, (&id, _)) in static_keys.iter().enumerate() {
-            let r2_secret = &round2_secret_packages[&id];
-            let round2_packages = &received_round2_packages[&id];
-            let (key_pkg, pubkey_pkg, _transcript, _cert) =
-                keys::cocktail_dkg::part3(r2_secret, round2_packages).unwrap();
-
-            assert_eq!(
-                pubkey_pkg.verifying_key().serialize().unwrap().as_slice(),
-                expected_group_public_key.as_slice(),
-                "participant {} group public key mismatch",
-                idx + 1
-            );
-            // JSON stores 56-byte raw scalars; our implementation uses 57-byte RFC 8032 format
-            // (little-endian scalar + trailing 0x00). Compare the first 56 bytes.
-            assert_eq!(
-                &key_pkg.signing_share().serialize().as_slice()[..56],
-                expected_shares[idx].as_slice(),
-                "participant {} secret share mismatch",
-                idx + 1
-            );
-            assert_eq!(
-                pubkey_pkg
-                    .verifying_shares()
-                    .get(&id)
-                    .unwrap()
-                    .serialize()
-                    .unwrap()
-                    .as_slice(),
-                expected_verification_shares[idx].as_slice(),
-                "participant {} verification share mismatch",
-                idx + 1
-            );
-        }
-
-        // Note: recovery is not tested here because the JSON ciphertexts use 56-byte raw scalar
-        // encryption while our implementation uses 57-byte RFC 8032 format, making them
-        // incompatible.
-    }
+    frost_core::tests::ciphersuite_generic::check_cocktail_dkg_test_vectors::<
+        Ed448Shake256,
+        _,
+        _,
+    >(
+        include_str!("helpers/cocktail-dkg-ed448-shake256.json"),
+        |seed, cs_id, t, n, p| CounterDrng::new(seed, cs_id, t, n, p),
+        false, // encrypted shares: 73 vs 72 byte format mismatch
+        false, // recovery: ciphertext format incompatible
+    );
 }
