+++ frost-p256/Cargo.toml
@@ -23,7 +23,6 @@ frost-core.workspace = true
 frost-rerandomized.workspace = true
 rand_core.workspace = true
 sha2 = { version = "0.10.2", default-features = false }
-xaes-256-gcm = "0.0.1-pre.0"
 
 [dev-dependencies]
 criterion.workspace = true
