+++ frost-secp256k1/Cargo.toml
@@ -23,7 +23,6 @@ frost-rerandomized.workspace = true
 k256 = { version = "0.13.0", features = ["arithmetic", "expose-field", "hash2curve"], default-features = false }
 rand_core.workspace = true
 sha2 = { version = "0.10.2", default-features = false }
-xaes-256-gcm = "0.0.1-pre.0"
 
 [dev-dependencies]
 criterion.workspace = true
